moved query interface and query parameters into protocols folder
