+++ metricflow/specs/where_filter_dimension.py
@@ -16,7 +16,7 @@ from typing_extensions import override
 
 from metricflow.errors.errors import InvalidQuerySyntax
 from metricflow.specs.column_assoc import ColumnAssociationResolver
-from metricflow.specs.query_interface import (
+from metricflow.protocols.query_interface import (
     QueryInterfaceDimension,
     QueryInterfaceDimensionFactory,
 )
